Remove /difficulty and /performance
The only benefit of /performance over /calculation is the tiny bit of
saved serialisation. Not worth the extra api and code complexity.

The only benefit of /difficulty over /calculation is avoiding the
performance calculation, but performance calc is over 1000x faster than
difficulty calculation. The only tangible benefit is when the
difficulties have already been calculated and are in the cache, but I
still don't think its worth the complexity of the extra api. Can add
back later if requested.